git-svn-id: https://subversion.assembla.com/svn/kbasm/cpgf/trunk@579 5ea4cdcf-77fc-412b-903d-47c6f9a32d48

+++ src/metautility/gmetabytearray.cpp
@@ -63,7 +63,7 @@ public:
 
     template <typename T>
     void write(T value, const GMetaVariadicParam * moreValues) {
-    	uint32_t valueCount = 1;
+    	size_t valueCount = 1;
     	if(moreValues != NULL) {
     		valueCount += moreValues->paramCount;
     	}
